+++ src/canister/src/storage/group_members_storage.rs
@@ -1,8 +1,11 @@
 use std::collections::HashMap;
 
-use super::storage_api::{
-    StaticStorageRef, Storage, StorageInsertableByKey, StorageQueryable, StorageUpdateable,
-    GROUP_MEMBERS, GROUP_MEMBERS_MEMORY_ID,
+use super::{
+    storage_api::{
+        StaticStorageRef, Storage, StorageInsertableByKey, StorageQueryable, StorageUpdateable,
+        GROUP_MEMBERS, GROUP_MEMBERS_MEMORY_ID,
+    },
+    ID_KIND_GROUP_MEMBERS,
 };
 use canister_types::models::member_collection::MemberCollection;
 use ic_stable_structures::memory_manager::MemoryId;
@@ -10,7 +13,7 @@ use ic_stable_structures::memory_manager::MemoryId;
 pub struct GroupMemberStore;
 
 impl Storage<u64, MemberCollection> for GroupMemberStore {
-    const NAME: &'static str = "group_members";
+    const NAME: &'static str = ID_KIND_GROUP_MEMBERS;
 
     fn storage() -> StaticStorageRef<u64, MemberCollection> {
         &GROUP_MEMBERS